requirements.txt

pyyaml
lxml
